move more stuff; rm unneeded trait
