+++ src/query/query.rs
@@ -1,5 +1,6 @@
-use crate::query::traversal::PairStorage;
-use crate::query::traversal::{OnePair, Present};
+use crate::query::traversal::PairMatchStore;
+pub(crate) use crate::query::traversal::PairMatcher;
+use crate::query::traversal::{ByRule, ByTag, OnePair, Present};
 use paste::paste;
 use pest::error::Error;
 use pest::iterators::{Pair, Pairs};
@@ -78,102 +79,6 @@ impl Query {
 }
 
 // TODO: I should reorganize this file, break it up into separate files etc
-
-pub trait PairMatcher {
-    fn matches(&self, pair: &Pair<Rule>) -> bool;
-
-    fn find_all_in(self, pairs: Pairs<Rule>) -> Vec<Pair<Rule>>
-    where
-        Self: Sized,
-    {
-        FindAll::new(self).find_in(pairs)
-    }
-}
-
-#[derive(Debug)]
-pub struct FindAll<'a, M>(M, Vec<Pair<'a, Rule>>);
-
-impl<'a, M> FindAll<'a, M> {
-    pub fn new(matcher: M) -> Self {
-        Self(matcher, Vec::new())
-    }
-}
-
-impl<'a, M> PairMatchStore<'a> for FindAll<'a, M>
-where
-    M: PairMatcher,
-{
-    type Output = Vec<Pair<'a, Rule>>;
-
-    fn match_and_store(&mut self, pair: Pair<'a, Rule>) -> Result<(), Pair<'a, Rule>> {
-        if self.0.matches(&pair) {
-            self.1.push(pair);
-            Ok(())
-        } else {
-            Err(pair)
-        }
-    }
-
-    fn get(self) -> Self::Output {
-        self.1
-    }
-}
-
-pub trait PairMatchStore<'a> {
-    type Output;
-
-    fn match_and_store(&mut self, pair: Pair<'a, Rule>) -> Result<(), Pair<'a, Rule>>;
-
-    fn get(self) -> Self::Output;
-
-    fn find_in(mut self, pairs: Pairs<'a, Rule>) -> Self::Output
-    where
-        Self: Sized,
-    {
-        fn build<'b>(me: &mut impl PairMatchStore<'b>, pairs: Pairs<'b, Rule>) {
-            for pair in pairs {
-                if let Err(unmatched) = me.match_and_store(pair) {
-                    build(me, unmatched.into_inner())
-                }
-            }
-        }
-        build(&mut self, pairs);
-        self.get()
-    }
-}
-
-#[derive(Debug)]
-pub struct ByRule(Rule);
-
-impl ByRule {
-    pub fn new(rule: Rule) -> Self {
-        Self(rule)
-    }
-}
-
-impl<'a> PairMatcher for ByRule {
-    fn matches(&self, pair: &Pair<Rule>) -> bool {
-        self.0 == pair.as_rule()
-    }
-}
-
-#[derive(Debug)]
-pub struct ByTag(&'static str);
-
-impl ByTag {
-    pub fn new(tag: &'static str) -> Self {
-        Self(tag)
-    }
-}
-
-impl PairMatcher for ByTag {
-    fn matches(&self, pair: &Pair<Rule>) -> bool {
-        match pair.as_node_tag() {
-            Some(t) => t == self.0,
-            None => false,
-        }
-    }
-}
 
 macro_rules! composite_finder {
     ($name:ident { $($elem:ident $result:ty : $finder:ident),+ $(,)? }) => {
