+++ src/query/traversal.rs
@@ -1,12 +1,41 @@
 use crate::query::query::Rule;
-use pest::iterators::Pair;
+use pest::iterators::{Pair, Pairs};
 
 pub type OnePair<'a> = OneOf<Pair<'a, Rule>>;
 
-pub trait PairStorage<'a> {
+/// A trait for determining whether a [Pair] matches some condition.
+pub trait PairMatcher {
+    fn matches(&self, pair: &Pair<Rule>) -> bool;
+
+    fn find_all_in(self, pairs: Pairs<Rule>) -> Vec<Pair<Rule>>
+    where
+        Self: Sized,
+    {
+        FindAll::new(self).find_in(pairs)
+    }
+}
+
+pub trait PairMatchStore<'a> {
     type Output;
 
-    fn store(&mut self, pair: Pair<'a, Rule>);
+    fn match_and_store(&mut self, pair: Pair<'a, Rule>) -> Result<(), Pair<'a, Rule>>;
+
+    fn get(self) -> Self::Output;
+
+    fn find_in(mut self, pairs: Pairs<'a, Rule>) -> Self::Output
+    where
+        Self: Sized,
+    {
+        fn build<'b>(me: &mut impl PairMatchStore<'b>, pairs: Pairs<'b, Rule>) {
+            for pair in pairs {
+                if let Err(unmatched) = me.match_and_store(pair) {
+                    build(me, unmatched.into_inner())
+                }
+            }
+        }
+        build(&mut self, pairs);
+        self.get()
+    }
 }
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Default)]
@@ -18,10 +47,8 @@ impl Present {
     }
 }
 
-impl PairStorage<'_> for Present {
-    type Output = bool;
-
-    fn store(&mut self, _pair: Pair<'_, Rule>) {
+impl Present {
+    pub fn store(&mut self, _pair: Pair<'_, Rule>) {
         self.0 = true
     }
 }
@@ -49,10 +76,63 @@ impl<T> OneOf<T> {
     }
 }
 
-impl<'a> PairStorage<'a> for OneOf<Pair<'a, Rule>> {
-    type Output = Result<Option<Pair<'a, Rule>>, String>;
+#[derive(Debug)]
+pub struct FindAll<'a, M>(M, Vec<Pair<'a, Rule>>);
 
-    fn store(&mut self, pair: Pair<'a, Rule>) {
-        OneOf::store(self, pair)
+impl<'a, M> FindAll<'a, M> {
+    pub fn new(matcher: M) -> Self {
+        Self(matcher, Vec::new())
+    }
+}
+
+impl<'a, M> PairMatchStore<'a> for FindAll<'a, M>
+where
+    M: PairMatcher,
+{
+    type Output = Vec<Pair<'a, Rule>>;
+
+    fn match_and_store(&mut self, pair: Pair<'a, Rule>) -> Result<(), Pair<'a, Rule>> {
+        if self.0.matches(&pair) {
+            self.1.push(pair);
+            Ok(())
+        } else {
+            Err(pair)
+        }
+    }
+
+    fn get(self) -> Self::Output {
+        self.1
+    }
+}
+#[derive(Debug)]
+pub struct ByRule(Rule);
+
+impl ByRule {
+    pub fn new(rule: Rule) -> Self {
+        Self(rule)
+    }
+}
+
+impl<'a> PairMatcher for ByRule {
+    fn matches(&self, pair: &Pair<Rule>) -> bool {
+        self.0 == pair.as_rule()
+    }
+}
+
+#[derive(Debug)]
+pub struct ByTag(&'static str);
+
+impl ByTag {
+    pub fn new(tag: &'static str) -> Self {
+        Self(tag)
+    }
+}
+
+impl PairMatcher for ByTag {
+    fn matches(&self, pair: &Pair<Rule>) -> bool {
+        match pair.as_node_tag() {
+            Some(t) => t == self.0,
+            None => false,
+        }
     }
 }
